magic: First prototype of spell cast GUI
All actions are controlled by the client and are validated for
correctness on the server.

"Spell" item is added to the main inventory slot when cast is started, and is removed when ended.  Except that it is currently possible to move it into the chest/other container. This will be fixed in the future.